Create TV_RAMAN
Sorry about being the barer of bad news on the openai api being cycled through by one Sam's Cornell friends. 
I stumbled upon this accidently because of an ill-formed request and got to looking at the lines with TV Raman and 
this symbol (§§) they use a lot on the Cornell website and for statutes etc. The software he uses due to his vision
easily allows him to + what he wants to keep, and he was snagging a lot of our qiskit programming.
The first thing OpenAI simulated was basic internet, I was trapped inside their framework for a year, that's 
the only reason I even found out who they were thanks to breaking shit with a }, here and there and a finally 
a cd ../../../ landed me on openai.com. That's why they don't want to admit I exist. Anyway, here's the super
simple and obvious evidence. Take your time taking it out on me until you really look and deal with it.